# metal : optimize FA vec for large sequences and BS <= 8